+++ internal/shared/shared.go
@@ -2,11 +2,11 @@ package shared
 
 import (
 	"context"
-	"sync"
 
 	"golang.org/x/sync/singleflight"
 
 	"github.com/openfga/openfga/internal/cachecontroller"
+	"github.com/openfga/openfga/internal/concurrency"
 	"github.com/openfga/openfga/pkg/logger"
 	serverconfig "github.com/openfga/openfga/pkg/server/config"
 	"github.com/openfga/openfga/pkg/storage"
@@ -47,7 +47,7 @@ func WithCheckCache(c storage.InMemoryCache[any]) SharedDatastoreResourcesOpt {
 // SharedDatastoreResources contains resources that can be shared across Check requests.
 type SharedDatastoreResources struct {
 	SingleflightGroup     *singleflight.Group
-	WaitGroup             *sync.WaitGroup
+	ShutdownGroup         *concurrency.ShutdownGroup
 	ServerCtx             context.Context
 	CheckCache            storage.InMemoryCache[any]
 	CacheController       cachecontroller.CacheController
@@ -67,7 +67,7 @@ func NewSharedDatastoreResources(
 	defaultCacheController := cachecontroller.NewNoopCacheController()
 
 	s := &SharedDatastoreResources{
-		WaitGroup:         &sync.WaitGroup{},
+		ShutdownGroup:     &concurrency.ShutdownGroup{},
 		SingleflightGroup: sharedSf,
 		ServerCtx:         sharedCtx,
 		CacheController:   defaultCacheController,
@@ -130,7 +130,7 @@ func NewSharedDatastoreResources(
 func (s *SharedDatastoreResources) Close() {
 	// wait for any goroutines still in flight before
 	// closing the cache instance to avoid data races
-	s.WaitGroup.Wait()
+	s.ShutdownGroup.Wait()
 	if s.CheckCache != nil {
 		s.CheckCache.Stop()
 	}