+++ pkg/storage/storagewrappers/cached_datastore_test.go
@@ -18,6 +18,7 @@ import (
 
 	openfgav1 "github.com/openfga/api/proto/openfga/v1"
 
+	"github.com/openfga/openfga/internal/concurrency"
 	"github.com/openfga/openfga/internal/mocks"
 	"github.com/openfga/openfga/pkg/logger"
 	"github.com/openfga/openfga/pkg/storage"
@@ -43,8 +44,8 @@ func TestFindInCache(t *testing.T) {
 	maxSize := 10
 	ttl := 5 * time.Hour
 	sf := &singleflight.Group{}
-	wg := &sync.WaitGroup{}
-	ds := NewCachedDatastore(ctx, mockDatastore, mockCache, maxSize, ttl, sf, wg)
+	sg := &concurrency.ShutdownGroup{}
+	ds := NewCachedDatastore(ctx, mockDatastore, mockCache, maxSize, ttl, sf, sg)
 
 	storeID := ulid.Make().String()
 	key := "key"
@@ -158,8 +159,6 @@ func TestReadStartingWithUser(t *testing.T) {
 	maxSize := 10
 	ttl := 5 * time.Hour
 	sf := &singleflight.Group{}
-	wg := &sync.WaitGroup{}
-	ds := NewCachedDatastore(ctx, mockDatastore, mockCache, maxSize, ttl, sf, wg)
 
 	storeID := ulid.Make().String()
 
@@ -226,6 +225,8 @@ func TestReadStartingWithUser(t *testing.T) {
 			mockCache.EXPECT().Delete(invalidEntityKeys[1]),
 		)
 
+		sg := &concurrency.ShutdownGroup{}
+		ds := NewCachedDatastore(ctx, mockDatastore, mockCache, maxSize, ttl, sf, sg)
 		iter, err := ds.ReadStartingWithUser(ctx, storeID, filter, options)
 		require.NoError(t, err)
 
@@ -247,7 +248,7 @@ func TestReadStartingWithUser(t *testing.T) {
 		iter.Stop() // has to be sync otherwise the assertion fails
 		i, ok := iter.(*cachedIterator)
 		require.True(t, ok)
-		i.wg.Wait()
+		i.sg.Wait()
 
 		if diff := cmp.Diff(tuples, actual, cmpOpts...); diff != "" {
 			t.Fatalf("mismatch (-want +got):\n%s", diff)
@@ -263,6 +264,8 @@ func TestReadStartingWithUser(t *testing.T) {
 				mockCache.EXPECT().Get(invalidEntityKeys[1]).Return(nil),
 			)
 
+			sg := &concurrency.ShutdownGroup{}
+			ds := NewCachedDatastore(ctx, mockDatastore, mockCache, maxSize, ttl, sf, sg)
 			iter, err := ds.ReadStartingWithUser(ctx, storeID, filter, options)
 			require.NoError(t, err)
 
@@ -308,6 +311,8 @@ func TestReadStartingWithUser(t *testing.T) {
 				mockCache.EXPECT().Get(invalidEntityKeysWithRelation[1]).Return(nil),
 			)
 
+			sg := &concurrency.ShutdownGroup{}
+			ds := NewCachedDatastore(ctx, mockDatastore, mockCache, maxSize, ttl, sf, sg)
 			iter, err := ds.ReadStartingWithUser(ctx, storeID, filterWithUserRelation, options)
 			require.NoError(t, err)
 
@@ -350,6 +355,8 @@ func TestReadStartingWithUser(t *testing.T) {
 			mockCache.EXPECT().Delete(invalidEntityKeys[1]),
 		)
 
+		sg := &concurrency.ShutdownGroup{}
+		ds := NewCachedDatastore(ctx, mockDatastore, mockCache, maxSize, ttl, sf, sg)
 		iter, err := ds.ReadStartingWithUser(ctx, storeID, filter, options)
 		require.NoError(t, err)
 
@@ -371,7 +378,7 @@ func TestReadStartingWithUser(t *testing.T) {
 		iter.Stop() // has to be sync otherwise the assertion fails
 		i, ok := iter.(*cachedIterator)
 		require.True(t, ok)
-		i.wg.Wait()
+		i.sg.Wait()
 
 		require.Empty(t, actual)
 	})
@@ -389,6 +396,8 @@ func TestReadStartingWithUser(t *testing.T) {
 				Return(storage.NewStaticTupleIterator(tuples), nil),
 		)
 
+		sg := &concurrency.ShutdownGroup{}
+		ds := NewCachedDatastore(ctx, mockDatastore, mockCache, maxSize, ttl, sf, sg)
 		iter, err := ds.ReadStartingWithUser(ctx, storeID, filter, opts)
 		require.NoError(t, err)
 		defer iter.Stop()
@@ -428,8 +437,6 @@ func TestReadUsersetTuples(t *testing.T) {
 	maxSize := 10
 	ttl := 5 * time.Hour
 	sf := &singleflight.Group{}
-	wg := &sync.WaitGroup{}
-	ds := NewCachedDatastore(ctx, mockDatastore, mockCache, maxSize, ttl, sf, wg)
 
 	storeID := ulid.Make().String()
 
@@ -490,6 +497,8 @@ func TestReadUsersetTuples(t *testing.T) {
 			mockCache.EXPECT().Delete(invalidEntityKey),
 		)
 
+		sg := &concurrency.ShutdownGroup{}
+		ds := NewCachedDatastore(ctx, mockDatastore, mockCache, maxSize, ttl, sf, sg)
 		iter, err := ds.ReadUsersetTuples(ctx, storeID, filter, options)
 		require.NoError(t, err)
 
@@ -511,7 +520,7 @@ func TestReadUsersetTuples(t *testing.T) {
 		iter.Stop() // has to be sync otherwise the assertion fails
 		i, ok := iter.(*cachedIterator)
 		require.True(t, ok)
-		i.wg.Wait()
+		i.sg.Wait()
 
 		if diff := cmp.Diff(tuples, actual, cmpOpts...); diff != "" {
 			t.Fatalf("mismatch (-want +got):\n%s", diff)
@@ -525,6 +534,8 @@ func TestReadUsersetTuples(t *testing.T) {
 			mockCache.EXPECT().Get(invalidEntityKey).Return(nil),
 		)
 
+		sg := &concurrency.ShutdownGroup{}
+		ds := NewCachedDatastore(ctx, mockDatastore, mockCache, maxSize, ttl, sf, sg)
 		iter, err := ds.ReadUsersetTuples(ctx, storeID, filter, options)
 		require.NoError(t, err)
 		defer iter.Stop()
@@ -564,6 +575,8 @@ func TestReadUsersetTuples(t *testing.T) {
 			mockCache.EXPECT().Delete(invalidEntityKey),
 		)
 
+		sg := &concurrency.ShutdownGroup{}
+		ds := NewCachedDatastore(ctx, mockDatastore, mockCache, maxSize, ttl, sf, sg)
 		iter, err := ds.ReadUsersetTuples(ctx, storeID, filter, options)
 		require.NoError(t, err)
 
@@ -585,7 +598,7 @@ func TestReadUsersetTuples(t *testing.T) {
 		iter.Stop() // has to be sync otherwise the assertion fails
 		i, ok := iter.(*cachedIterator)
 		require.True(t, ok)
-		i.wg.Wait()
+		i.sg.Wait()
 
 		require.Empty(t, actual)
 	})
@@ -603,6 +616,8 @@ func TestReadUsersetTuples(t *testing.T) {
 				Return(storage.NewStaticTupleIterator(tuples), nil),
 		)
 
+		sg := &concurrency.ShutdownGroup{}
+		ds := NewCachedDatastore(ctx, mockDatastore, mockCache, maxSize, ttl, sf, sg)
 		iter, err := ds.ReadUsersetTuples(ctx, storeID, filter, opts)
 		require.NoError(t, err)
 		defer iter.Stop()
@@ -643,8 +658,6 @@ func TestRead(t *testing.T) {
 	maxSize := 10
 	ttl := 5 * time.Hour
 	sf := &singleflight.Group{}
-	wg := &sync.WaitGroup{}
-	ds := NewCachedDatastore(ctx, mockDatastore, mockCache, maxSize, ttl, sf, wg)
 
 	storeID := ulid.Make().String()
 
@@ -698,6 +711,8 @@ func TestRead(t *testing.T) {
 			mockCache.EXPECT().Delete(invalidEntityKey),
 		)
 
+		sg := &concurrency.ShutdownGroup{}
+		ds := NewCachedDatastore(ctx, mockDatastore, mockCache, maxSize, ttl, sf, sg)
 		iter, err := ds.Read(ctx, storeID, filter, storage.ReadOptions{})
 		require.NoError(t, err)
 
@@ -719,7 +734,7 @@ func TestRead(t *testing.T) {
 		iter.Stop() // has to be sync otherwise the assertion fails
 		i, ok := iter.(*cachedIterator)
 		require.True(t, ok)
-		i.wg.Wait()
+		i.sg.Wait()
 
 		if diff := cmp.Diff(tuples, actual, cmpOpts...); diff != "" {
 			t.Fatalf("mismatch (-want +got):\n%s", diff)
@@ -733,6 +748,8 @@ func TestRead(t *testing.T) {
 			mockCache.EXPECT().Get(invalidEntityKey).Return(nil),
 		)
 
+		sg := &concurrency.ShutdownGroup{}
+		ds := NewCachedDatastore(ctx, mockDatastore, mockCache, maxSize, ttl, sf, sg)
 		iter, err := ds.Read(ctx, storeID, filter, storage.ReadOptions{})
 		require.NoError(t, err)
 		defer iter.Stop()
@@ -772,6 +789,8 @@ func TestRead(t *testing.T) {
 			mockCache.EXPECT().Delete(invalidEntityKey),
 		)
 
+		sg := &concurrency.ShutdownGroup{}
+		ds := NewCachedDatastore(ctx, mockDatastore, mockCache, maxSize, ttl, sf, sg)
 		iter, err := ds.Read(ctx, storeID, filter, storage.ReadOptions{})
 		require.NoError(t, err)
 
@@ -793,7 +812,7 @@ func TestRead(t *testing.T) {
 		iter.Stop() // has to be sync otherwise the assertion fails
 		i, ok := iter.(*cachedIterator)
 		require.True(t, ok)
-		i.wg.Wait()
+		i.sg.Wait()
 
 		require.Empty(t, actual)
 	})
@@ -811,6 +830,8 @@ func TestRead(t *testing.T) {
 				Return(storage.NewStaticTupleIterator(tuples), nil),
 		)
 
+		sg := &concurrency.ShutdownGroup{}
+		ds := NewCachedDatastore(ctx, mockDatastore, mockCache, maxSize, ttl, sf, sg)
 		iter, err := ds.Read(ctx, storeID, filter, opts)
 		require.NoError(t, err)
 		defer iter.Stop()
@@ -848,6 +869,8 @@ func TestRead(t *testing.T) {
 				Return(storage.NewStaticTupleIterator(tuples), nil),
 		)
 
+		sg := &concurrency.ShutdownGroup{}
+		ds := NewCachedDatastore(ctx, mockDatastore, mockCache, maxSize, ttl, sf, sg)
 		iter, err := ds.Read(ctx, storeID, invalidObjectFilter, storage.ReadOptions{})
 		require.NoError(t, err)
 		defer iter.Stop()
@@ -887,8 +910,8 @@ func TestDatastoreIteratorError(t *testing.T) {
 	maxSize := 10
 	ttl := 5 * time.Hour
 	sf := &singleflight.Group{}
-	wg := &sync.WaitGroup{}
-	ds := NewCachedDatastore(ctx, mockDatastore, mockCache, maxSize, ttl, sf, wg)
+	sg := &concurrency.ShutdownGroup{}
+	ds := NewCachedDatastore(ctx, mockDatastore, mockCache, maxSize, ttl, sf, sg)
 
 	storeID := ulid.Make().String()
 
@@ -973,7 +996,7 @@ func TestCachedIterator(t *testing.T) {
 			maxResultSize:     maxCacheSize,
 			ttl:               ttl,
 			sf:                &singleflight.Group{},
-			wg:                &sync.WaitGroup{},
+			sg:                &concurrency.ShutdownGroup{},
 			objectType:        "",
 			objectID:          "",
 			relation:          "",
@@ -1014,7 +1037,7 @@ func TestCachedIterator(t *testing.T) {
 			maxResultSize:     maxCacheSize,
 			ttl:               ttl,
 			sf:                &singleflight.Group{},
-			wg:                &sync.WaitGroup{},
+			sg:                &concurrency.ShutdownGroup{},
 			objectType:        "",
 			objectID:          "",
 			relation:          "",
@@ -1049,7 +1072,7 @@ func TestCachedIterator(t *testing.T) {
 			maxResultSize:     maxCacheSize,
 			ttl:               ttl,
 			sf:                &singleflight.Group{},
-			wg:                &sync.WaitGroup{},
+			sg:                &concurrency.ShutdownGroup{},
 			objectType:        "",
 			objectID:          "",
 			relation:          "",
@@ -1077,7 +1100,7 @@ func TestCachedIterator(t *testing.T) {
 		}
 
 		iter.Stop()
-		iter.wg.Wait()
+		iter.sg.Wait()
 
 		cachedResults := cache.Get(cacheKey)
 		require.Nil(t, cachedResults)
@@ -1106,7 +1129,7 @@ func TestCachedIterator(t *testing.T) {
 			maxResultSize:     maxCacheSize,
 			ttl:               ttl,
 			sf:                &singleflight.Group{},
-			wg:                &sync.WaitGroup{},
+			sg:                &concurrency.ShutdownGroup{},
 			objectType:        "",
 			objectID:          "",
 			relation:          "",
@@ -1115,7 +1138,7 @@ func TestCachedIterator(t *testing.T) {
 		}
 
 		iter.Stop()
-		iter.wg.Wait()
+		iter.sg.Wait()
 
 		cachedResults := cache.Get(cacheKey)
 		require.Nil(t, cachedResults)
@@ -1144,7 +1167,7 @@ func TestCachedIterator(t *testing.T) {
 			maxResultSize:     maxCacheSize,
 			ttl:               ttl,
 			sf:                &singleflight.Group{},
-			wg:                &sync.WaitGroup{},
+			sg:                &concurrency.ShutdownGroup{},
 			objectType:        "",
 			objectID:          "",
 			relation:          "",
@@ -1172,7 +1195,7 @@ func TestCachedIterator(t *testing.T) {
 		}
 
 		iter.Stop()
-		iter.wg.Wait()
+		iter.sg.Wait()
 
 		cachedResults := cache.Get(cacheKey)
 		require.NotNil(t, cachedResults)
@@ -1207,7 +1230,7 @@ func TestCachedIterator(t *testing.T) {
 			maxResultSize:     maxCacheSize,
 			ttl:               ttl,
 			sf:                &singleflight.Group{},
-			wg:                &sync.WaitGroup{},
+			sg:                &concurrency.ShutdownGroup{},
 			objectType:        "",
 			objectID:          "",
 			relation:          "",
@@ -1216,7 +1239,7 @@ func TestCachedIterator(t *testing.T) {
 		}
 
 		iter.Stop()
-		iter.wg.Wait()
+		iter.sg.Wait()
 
 		cachedResults := cache.Get(cacheKey)
 		require.NotNil(t, cachedResults)
@@ -1251,7 +1274,7 @@ func TestCachedIterator(t *testing.T) {
 			maxResultSize:     maxCacheSize,
 			ttl:               ttl,
 			sf:                &singleflight.Group{},
-			wg:                &sync.WaitGroup{},
+			sg:                &concurrency.ShutdownGroup{},
 			objectType:        "",
 			objectID:          "",
 			relation:          "",
@@ -1266,7 +1289,7 @@ func TestCachedIterator(t *testing.T) {
 		require.ErrorIs(t, err, context.Canceled)
 
 		iter.Stop()
-		iter.wg.Wait()
+		iter.sg.Wait()
 
 		cachedResults := cache.Get(cacheKey)
 		require.NotNil(t, cachedResults)
@@ -1317,7 +1340,7 @@ func TestCachedIterator(t *testing.T) {
 			maxResultSize:     maxCacheSize,
 			ttl:               ttl,
 			sf:                &singleflight.Group{},
-			wg:                &sync.WaitGroup{},
+			sg:                &concurrency.ShutdownGroup{},
 			objectType:        "",
 			objectID:          "",
 			relation:          "",
@@ -1334,7 +1357,7 @@ func TestCachedIterator(t *testing.T) {
 		}()
 
 		wg.Wait()
-		iter.wg.Wait()
+		iter.sg.Wait()
 
 		require.Zero(t, mockedIter.nextCalled)
 		require.Nil(t, iter.tuples)
@@ -1378,7 +1401,7 @@ func TestCachedIterator(t *testing.T) {
 			ttl:               ttl,
 			initializedAt:     time.Now(),
 			sf:                &singleflight.Group{},
-			wg:                &sync.WaitGroup{},
+			sg:                &concurrency.ShutdownGroup{},
 			objectType:        "",
 			objectID:          "",
 			relation:          "",
@@ -1395,7 +1418,7 @@ func TestCachedIterator(t *testing.T) {
 		}()
 
 		wg.Wait()
-		iter.wg.Wait()
+		iter.sg.Wait()
 
 		require.Zero(t, mockedIter.nextCalled)
 		require.Nil(t, iter.tuples)
@@ -1438,7 +1461,7 @@ func TestCachedIterator(t *testing.T) {
 				maxResultSize:     maxCacheSize,
 				ttl:               ttl,
 				sf:                sf,
-				wg:                &sync.WaitGroup{},
+				sg:                &concurrency.ShutdownGroup{},
 				objectType:        "",
 				objectID:          "",
 				relation:          "",
@@ -1462,7 +1485,7 @@ func TestCachedIterator(t *testing.T) {
 				maxResultSize:     maxCacheSize,
 				ttl:               ttl,
 				sf:                sf,
-				wg:                &sync.WaitGroup{},
+				sg:                &concurrency.ShutdownGroup{},
 				objectType:        "",
 				objectID:          "",
 				relation:          "",
@@ -1476,14 +1499,14 @@ func TestCachedIterator(t *testing.T) {
 				defer wg.Done()
 
 				iter1.Stop()
-				iter1.wg.Wait()
+				iter1.sg.Wait()
 			}()
 
 			go func() {
 				defer wg.Done()
 
 				iter2.Stop()
-				iter2.wg.Wait()
+				iter2.sg.Wait()
 			}()
 
 			wg.Wait()